# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Vedetta NVR API
  version: "0.1.0"
  description: API for managing cameras, events, recordings, and object detection in the Vedetta Network Video Recorder.
servers:
  - url: /
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/system/recompress/trigger:
    post:
      operationId: triggerRecompression
      summary: Trigger manual recompression
      tags: [health]
      responses:
        "202":
          description: Recompression started
        "409":
          description: Recompression already in progress
          content:
            text/plain:
              schema:
                type: string
...
  /api/cameras/{name}:
    get:
      operationId: getCamera
      summary: Get camera detail
      tags: [cameras]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Camera detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CameraDetail"
        "404":
          description: Camera not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CameraDetail:
      type: object
      required: [name, online, has_motion, degraded, ptz, zone_count, recording]
      properties:
        name:
          type: string
        online:
          type: boolean
        has_motion:
          type: boolean
        degraded:
          type: boolean
        degraded_reason:
          type: string
        last_frame:
          type: string
          format: date-time
        ptz:
          type: boolean
        zone_count:
          type: integer
        recording:
          type: boolean
        source_fps:
          type: number
          format: float
          description: Rolling-window source frame rate measured at the detect consumer (frames per second).
...